python/main_project/cryptonite/requirements.txt
pynacl
rich
pefile